requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
asyncpg==0.30.0
psycopg2-binary==2.9.11
pydantic==2.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
cloudinary==1.44.1
aiohttp==3.9.1
aiogram==3.7.0
aiofiles==23.2.1
